pyproject.toml
[project]
name = "biodcase-2025-task1"
version = "0.1.0"
description = "BioDCASE 2025 Task 1"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "joblib==1.4.2",
  "librosa==0.10.2.post1",
  "numpy==1.26.4",
  "pandas==2.2.3",
  "PyYAML==6.0.2",
  "torch==2.6.0",
  "torchaudio==2.6.0",
  "tqdm==4.67.1",
]

